+++ src/count_min_sketch.h
@@ -16,6 +16,7 @@ typedef uint64_t* (*cms_hash_function) (int num_hashes, char *key);
 typedef struct {
     int depth;
     int width;
+    long elements_added;
     cms_hash_function hash_function;
     int** bins;
 }  CountMinSketch, count_min_sketch;
@@ -24,6 +25,11 @@ typedef struct {
 int cms_init_alt(CountMinSketch *cms, int width, int depth, cms_hash_function hash_function);
 static int __inline__  cms_init(CountMinSketch *cms, int width, int depth) {
     return cms_init_alt(cms, width, depth, NULL);
+}
+
+int cms_init_optimal_alt(CountMinSketch *cms, float error_rate, float confidence, cms_hash_function hash_function);
+static int __inline__  cms_init_optimal(CountMinSketch *cms, float error_rate, float confidence) {
+    return cms_init_optimal_alt(cms, error_rate, confidence, NULL);
 }
 
 int cms_destroy(CountMinSketch *cms);
@@ -35,7 +41,9 @@ int cms_add(CountMinSketch *cms, char* key);
 int cms_remove(CountMinSketch *cms, char* key);
 
 int cms_check(CountMinSketch *cms, char* key);
-int cms_check_min(CountMinSketch *cms, char* key);
+static int __inline__ cms_check_min(CountMinSketch *cms, char* key) {
+    return cms_check(cms, key);
+}
 
 int cms_check_mean(CountMinSketch *cms, char* key);